Update bindgen to 0.59

+++ src/lib.rs
@@ -586,7 +586,7 @@ pub fn create_shader_module(spv_data: &[u8]) -> Result<ShaderModule, &'static st
     let mut module: ffi::SpvReflectShaderModule = unsafe { std::mem::zeroed() };
     let result: ffi::SpvReflectResult = unsafe {
         ffi::spvReflectCreateShaderModule(
-            spv_data.len(),
+            spv_data.len() as u64,
             spv_data.as_ptr() as *const std::os::raw::c_void,
             &mut module,
         )
